*Update Sorcerer Zephyr skill behavior: -Gives immunity to all physical attacks except normal melee hits. -Gives immunity to single target magic attacks. -Zehpyr does not affect Ventus. *Update Soulinker Kaite skill behavior. *Recoded cashshop_list function. *Optimize flag attacks in battle.c.

+++ src/map/battle.h
@@ -46,9 +46,9 @@ struct Damage {
 #ifdef RENEWAL
 	int statusAtk, statusAtk2, weaponAtk, weaponAtk2, equipAtk, equipAtk2, masteryAtk, masteryAtk2;
 #endif
-	int64 damage,damage2; // Right, left dmg
-	int type,div_; // Chk clif_damage for type @TODO add an enum ? ; nb of hit
-	int amotion,dmotion;
+	uint32 damage,damage2; // Right, left dmg
+	int type, div_; // Chk clif_damage for type @TODO add an enum ? ; nb of hit
+	int amotion, dmotion;
 	int blewcount; // Nb of knockback
 	int flag; // Chk BF_* flag, (enum below)
 	int miscflag; //
